Make scripts work

+++ kubernetes/on_openstack_cli.sh
@@ -1,25 +1,60 @@
 #!/bin/bash
 
-set -x
-ips=(192.168.24 192.168.29.22 192.168.29.29 192.168.29.23)
+function validateIP() {
+  if [[ $1 =~ ^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$ ]]; then
+    echo 0
+  else
+    echo 1
+  fi
+}
 
-# copying all the ips to the ubuntu firts ssh
-echo "${ips[@]}" > /tmp/ips
-scp -i kp- -oStrictHostKeyChecking=no -P 10243 /tmp/ips ubuntu@openstack.math.unipd.it:/tmp/ips
-
-echo ${ips[@]}
-rm .ssh/known_hosts
-ssh -i kp- -oStrictHostKeyChecking=no -p 10243 ubuntu@openstack.math.unipd.it "bash -x <(curl -s https://raw.githubusercontent.com/Augugrumi/init-script/e09256e9939cedfe7b3f17377b747f325886451c/kubernetes/install_kubectl_repo.sh)"
-
-# create etc/hosts
-for i in ${ips[@]}; do
-  echo "$i $(nova list | grep $i | cut -d"|" -f3 | tr -d '[:space:]')" > /tmp/hosts
+ips=()
+while getopts ":a:" opt; do
+        case $opt in
+            a)
+		valid=$(validateIP $OPTARG)
+                if [ "$valid" -eq 0 ]; then
+                  ips+=("$OPTARG")
+                else
+                  msg err "Invalid IP address: -$OPTARG" >&2
+                  toExit=1
+                fi
+                ;;
+            \?)
+                msg err "Invalid option: -$OPTARG" >&2
+                toExit=1
+                ;;
+            :)
+                msg err "Option -$OPTARG requires an argument." >&2
+                toExit=1
+                ;;
+        esac
 done
 
-# copying hosts to first ssh
-scp -i kp- -oStrictHostKeyChecking=no -P 10243 /tmp/hosts ubuntu@openstack.math.unipd.it:/home/ubuntu
+if [ "$toExit" -ne 0 ]; then
+    echo "Shit's on fire bro"
+  else
+    # copying all the ips to the ubuntu firts ssh
+    echo "${ips[@]}" > /tmp/ips
+    scp -i kp- -oStrictHostKeyChecking=no -P 10243 /tmp/ips ubuntu@openstack.math.unipd.it:/tmp/ips
 
-# wait that all host are ready
-ssh -oStrictHostKeyChecking=no -i kp- -p 10243 ubuntu@openstack.math.unipd.it "mapfile ips < /tmp/ips; echo $(pwd); rc=1; for i in ${ips[@]}; do while [[ $rc -ne 0 ]]; do ping $i -c1; rc=$?; done; rc=1; done; for i in `${ips[@]}`; do scp -i kp- -oStrictHostKeyChecking=no /tmp/hosts ubuntu@$i:/etc/hosts; done" 
+    echo ${ips[@]}
+    rm .ssh/known_hosts
+    ssh -i kp- -oStrictHostKeyChecking=no -p 10243 ubuntu@openstack.math.unipd.it "bash -x <(curl -s https://raw.githubusercontent.com/Augugrumi/init-script/5ac0bc9aa79f9d8f10f1af3a7529af62f0d362d7/kubernetes/install_kubectl_repo.sh)"
 
-ssh -i kp- -oStrictHostKeyChecking=no -p 10243 ubuntu@openstack.math.unipd.it "ssh ubuntu@192.168.29.22 -oStrictHostKeyChecking=no -i kp- \"bash <(wget -qO- https://www.dropbox.com/s/qjvint34vythj7x/p.sh?dl=1)\" "
+    echo "" > /tmp/hosts
+    # create etc/hosts
+    for i in ${ips[@]}; do
+      echo "$i $(nova list | grep $i | cut -d"|" -f3 | tr -d '[:space:]')" >> /tmp/hosts
+    done
+
+    # copying hosts to first ssh
+    scp -i kp- -oStrictHostKeyChecking=no -P 10243 /tmp/hosts ubuntu@openstack.math.unipd.it:/tmp/hosts
+
+    sleep 30
+
+    # wait that all host are ready
+    ssh -oStrictHostKeyChecking=no -i kp- -p 10243 ubuntu@openstack.math.unipd.it "bash -x <(curl -s https://raw.githubusercontent.com/Augugrumi/init-script/5ac0bc9aa79f9d8f10f1af3a7529af62f0d362d7/kubernetes/wait_and_copy_hosts.sh)"
+
+    ssh -i kp- -oStrictHostKeyChecking=no -p 10243 ubuntu@openstack.math.unipd.it "ssh ubuntu@${ips[0]} -oStrictHostKeyChecking=no -i kp- \"bash -x <(curl -s https://raw.githubusercontent.com/Augugrumi/init-script/52f8ef8f61e9f37069ab2d90a9619bb81622c1f6/kubernetes/on_master.sh)\""
+fi
